pyproject.toml
[project]
name = "stock-ticker-analyzer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "numpy",
    "pandas",
    "openai>=1.0",
    "databricks-sdk",
]
